pyserver/requirements.txt
# 核心依赖
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.8.1
pydantic>=2.10.6
python-dotenv>=1.0.0
structlog>=25.1.0

# MQTT通信
paho-mqtt>=1.6.1

# 异步支持（可选）
anyio>=4.0.0

pyserial>=3.5


# 代码质量
black>=24.0.0
flake8>=7.0.0
mypy>=1.15.0


# 开发工具
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-multipart>=0.0.9

watchfiles>=0.19.0
ipython>=8.12.0
colorama>=0.4.6